refactor(web): share CAT header pickers at project level
